move weekly crates
